+++ src/ewaluacja2021/const.py
@@ -7,3 +7,14 @@ DEC2INT = 10000
 
 MAX_SLOT_AUT = 4
 MAX_SLOT_MONO = 2
+
+IMPORT_MAKSYMALNYCH_SLOTOW_COLUMNS = [
+    "Stopień / Tytuł",
+    "Nazwisko",
+    "Imię",
+    "ORCID",
+    "Dyscyplina",
+    "Uwzględniony w ewaluacji",
+    "maksymalna suma udziałów jednostkowych - wszystkie dyscypliny",
+    "maksymalna suma udziałów jednostkowych - monografie",
+]