+++ src/ewaluacja2021/tests/utils.py
@@ -1,0 +1,5 @@
+import os
+
+
+def curdir(fn):
+    return os.path.join(os.path.dirname(__name__), fn)